requirements/main.txt
hydra-core==1.0.6
numpy==1.20.3
torch==1.8.1
omegaconf==2.0.6
setuptools==52.0.0
gym==0.18.3
tqdm==4.59.0
tensorboard==2.5.0